cmd/go-cache-plugin: move setup code to its own file
