+++ cmd/go-cache-plugin/commands.go
@@ -21,16 +21,12 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/aws/aws-sdk-go-v2/config"
-	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/creachadair/command"
 	"github.com/creachadair/gocache"
-	"github.com/creachadair/gocache/cachedir"
+	"github.com/creachadair/mhttp/proxyconn"
 	"github.com/creachadair/taskgroup"
 	"github.com/goproxy/goproxy"
-	"github.com/tailscale/go-cache-plugin/internal/s3util"
 	"github.com/tailscale/go-cache-plugin/revproxy"
-	"github.com/tailscale/go-cache-plugin/s3cache"
 	"github.com/tailscale/go-cache-plugin/s3proxy"
 	"tailscale.com/tsweb"
 )
@@ -47,63 +43,6 @@ var flags struct {
 	Expiration    time.Duration `flag:"expiry,default=$GOCACHE_EXPIRY,Cache expiration period (optional)"`
 	Verbose       bool          `flag:"v,default=$GOCACHE_VERBOSE,Enable verbose logging"`
 	DebugLog      bool          `flag:"debug,default=$GOCACHE_DEBUG,Enable detailed per-request debug logging (noisy)"`
-}
-
-func initCacheServer(env *command.Env) (*gocache.Server, *s3util.Client, error) {
-	switch {
-	case flags.CacheDir == "":
-		return nil, nil, env.Usagef("you must provide a --cache-dir")
-	case flags.S3Bucket == "":
-		return nil, nil, env.Usagef("you must provide an S3 --bucket name")
-	}
-	region, err := getBucketRegion(env.Context(), flags.S3Bucket)
-	if err != nil {
-		return nil, nil, env.Usagef("you must provide an S3 --region name")
-	}
-
-	dir, err := cachedir.New(flags.CacheDir)
-	if err != nil {
-		return nil, nil, fmt.Errorf("create local cache: %w", err)
-	}
-
-	cfg, err := config.LoadDefaultConfig(env.Context(), config.WithRegion(region))
-	if err != nil {
-		return nil, nil, fmt.Errorf("laod AWS config: %w", err)
-	}
-
-	vprintf("local cache directory: %s", flags.CacheDir)
-	vprintf("S3 cache bucket %q (%s)", flags.S3Bucket, region)
-	client := &s3util.Client{
-		Client: s3.NewFromConfig(cfg),
-		Bucket: flags.S3Bucket,
-	}
-	cache := &s3cache.Cache{
-		Local:             dir,
-		S3Client:          client,
-		KeyPrefix:         flags.KeyPrefix,
-		MinUploadSize:     flags.MinUploadSize,
-		UploadConcurrency: flags.S3Concurrency,
-	}
-	cache.SetMetrics(env.Context(), expvar.NewMap("gocache_host"))
-
-	close := cache.Close
-	if flags.Expiration > 0 {
-		dirClose := dir.Cleanup(flags.Expiration)
-		close = func(ctx context.Context) error {
-			return errors.Join(cache.Close(ctx), dirClose(ctx))
-		}
-	}
-	s := &gocache.Server{
-		Get:         cache.Get,
-		Put:         cache.Put,
-		Close:       close,
-		SetMetrics:  cache.SetMetrics,
-		MaxRequests: flags.Concurrency,
-		Logf:        vprintf,
-		LogRequests: flags.DebugLog,
-	}
-	expvar.Publish("gocache_server", s.Metrics().Get("server"))
-	return s, client, nil
 }
 
 // runDirect runs a cache communicating on stdin/stdout, for use as a direct
